Fixed MORE integer to double problems for Vishnu

+++ BackEnd/src/main/java/model/Availability.java
@@ -5,11 +5,11 @@ public class Availability {
     private int eventId;
     private int personId;
     private int availabilityId;
-    private int startTime;
-    private int endTime;
+    private double startTime;
+    private double endTime;
     private int dow;
 
-    public Availability(int eId, int pId, int st, int et, int dow) {
+    public Availability(int eId, int pId, double st, double et, int dow) {
         this.eventId=eId;
         this.personId=pId;
         this.startTime=st;
@@ -41,19 +41,19 @@ public class Availability {
         this.availabilityId = availabilityId;
     }
 
-    public int getStartTime() {
+    public double getStartTime() {
         return startTime;
     }
 
-    public void setStartTime(int startTime) {
+    public void setStartTime(double startTime) {
         this.startTime = startTime;
     }
 
-    public int getEndTime() {
+    public double getEndTime() {
         return endTime;
     }
 
-    public void setEndTime(int endTime) {
+    public void setEndTime(double endTime) {
         this.endTime = endTime;
     }
